+++ bigquery/src/lib.rs
@@ -4,5 +4,5 @@ pub mod arrow;
 pub mod client;
 pub mod grpc;
 pub mod http;
-pub mod iterator;
-pub mod value;
+pub mod storage;
+pub mod query;
